feat: nonce_validation_service and updated postman collection
