+++ agent_issuance/Cargo.toml
@@ -11,6 +11,7 @@ agent_secret_manager = { path = "../agent_secret_manager" }
 
 anyhow.workspace = true
 async-trait.workspace = true
+base64.workspace = true
 cqrs-es.workspace = true
 chrono.workspace = true
 types-ob-v3 = { git = "https://github.com/impierce/digital-credential-data-models", rev = "9f16c27" }
