+++ agent_issuance/src/application/mod.rs
@@ -1,2 +1,3 @@
 pub mod access_token_validation_service;
+pub mod nonce_validation_service;
 pub mod policies;
